pyproject.toml
[project]
name = "fig"
version = "0.1.0"
authors = [
    { name = "pythoninthegrass", email = "4097471+pythoninthegrass@users.noreply.github.com" }
]
license = { file = "LICENSE" }

requires-python = ">=3.12,<3.13"

dependencies = [
    "pictex>=1.1.1",
    "pyfiglet>=0.8.0",
    "python-decouple>=3.8",
]

[project.optional-dependencies]
dev = [
    "argcomplete<4.0.0,>=3.5.0",
    "deptry<1.0.0,>=0.23.0",
    "icecream<3.0.0,>=2.1.3",
    "ipython<9.0.0,>=8.27.0",
    "mypy<2.0.0,>=1.14.1",
    "pyclean<4.0.0,>=3.0.0",
    "pytest-asyncio<1.0.0,>=0.25.2",
    "pytest-cov<7.0.0,>=6.0.0",
    "pytest<9.0.0,>=8.3.4",
    "rich<14.0.0,>=13.8.1",
    "ruff>=0.9.5",
]
test = [
    "coverage<8.0.0,>=7.6.1",
    "hypothesis[cli]<7.0.0,>=6.112.1",
    "pytest<9.0.0,>=8.3.3",
    "pytest-asyncio<1.0.0,>=0.24.0",
    "pytest-cov>=6.1.1",
    "pytest-datafiles<4.0.0,>=3.0.0",
    "pytest-xdist<4.0.0,>=3.6.1",
]

[tool.deptry]
# DEP003: transitive deps
ignore = [
    "DEP003"
]

[tool.deptry.per_rule_ignores]
# DEP002: not used in codebase (excluding dev deps)
DEP002 = [
    "deptry",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "uvicorn"
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
asyncio_mode = "auto"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
    "benchmark: marks performance benchmark tests"
]

[tool.coverage.run]
source = ["{{ cookiecutter.python_package_import_name }}"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/alembic/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
    "..."
]
